`args`/`kwargs` in rels; Closes #34
This is not a breaking update, but it does represent a major update to
the style of ConstraintHg, in allowing relations (`rel`, `via`,
`index_via`, etc.) to be expressed without the buffer `*args`, `**kwargs`
arguments. Before this, all relations had to be able to handle any number
of arguments. This made it impractical to use `*args` or `**kwargs` within
a method, and caused several bugs when the boilerplate `**kwargs` wasn't
added.

The fix for this issue is twofold:
1. Transform all passed values for an edge (the sources) into a dictionary,
even if that means generic labels. The construction of this dictionary is
such that it tries to fill every keyword argument required by a relation
(`via`, `rel`, etc.). This was accomplished primarily by editing
`identify_labeled_source_nodes`.
2. When calling a relation, filter the dict of source nodes by which
values are needed for the relation. This is accomplished by adding a new
method `filtered_call`. The method adheres to python argument positioning.
It is used in every call for a relation, including `index_via` and with
super nodes.

Finally, tests have been added to ensure basic functionality.

This change should not break any workflows, as relations with *args, **kwargs
are still acceptable and processed normally.